Update memory loss condition

+++ Assets/Scripts/Damager.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class Damager : MonoBehaviour
 {
-    public int damageDealt;
+    public float damageDealt;
     public DamageType[] damageTypes = new DamageType[] { DamageType.Physical };
 
-    public int DoDamage(Health target)
+    public float DoDamage(Health target)
     {
         return target.Damage(damageDealt, damageTypes);
     }
